api-server/requirements.txt
Django==1.5.4
MySQL-python==1.2.3
Pillow==2.3.0
djangorestframework==2.3.13
mutagen==1.22
pyOpenSSL==0.13.1
pycrypto==2.6.1
sorl-thumbnail==11.12
tweepy==2.2
